requirements.txt
# ConvoLens Project Dependencies

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# API and web requests
anthropic>=0.18.0
requests>=2.31.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web interface
gradio>=4.0.0

# Machine learning utilities
scikit-learn>=1.3.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Environment management
python-dotenv>=1.0.0

# Text processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Progress bars
tqdm>=4.65.0

# Excel support (for annotation workflow)
openpyxl>=3.1.0
